Refactor with helper file
